pyproject.toml
[project]
name = "pet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "aiofiles>=24.1.0",
    "aiosmtplib>=4.0.2",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "mixer>=7.2.2",
    "orjson>=3.11.3",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[all,email]>=2.11.7",
    "pyjwt[crypt,crypto]>=2.10.1",
    "pytest-async>=0.1.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "python-multipart>=0.0.20",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn[all]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-async>=0.1.1",
    "pytest-cov>=6.2.1",
]


[tool.pytest.ini_options]
asyncio_mode = 'auto'
asyncio_default_test_loop_scope = 'session'
asyncio_default_fixture_loop_scope = 'session'
